requirements.txt
Flask==2.2.2
click==8.1.3
Werkzeug==2.2.2
setuptools==60.2.0
wfastcgi==3.0.0
Flask-SocketIO==5.3.2
eventlet==0.25.2
daphne==4.0.0